PJKT-79: add one more test for each service. Plus fixes.

+++ src/main/java/com/github/StudentsDreamTeam/dto/ItemDTO.java
@@ -10,8 +10,8 @@ public record ItemDTO(
         String name,
         String description,
         String rarity,
-        Long xpMultiplier,
-        Long currencyMultiplier,
+        Float xpMultiplier,
+        Float currencyMultiplier,
         Long duration,
         Long cost
 ) {
